Policy: detect config.h and squid.h include problems
We currently have a policy that config.h MUST be included first so as to
pull in the portability definitions early. It MAY be included via the
legacy squid.h at present.

This alteration to the source maintenance script validates that each .c
and .cc file in the sources includes config.h or squid.h first in its
include order. Also that each .h and .cci do not include config.h which
is a double include with enforced .c/.cc requirement.

 * an ERROR: line is produces for each violating file
 * as yet the maintenance run is not blocked so as to catch as many
   errors as possible in one run
 * detection only, as yet no code alterations are performed by this script
 * FORMAT: informative lines are silenced to make ERROR: more visible


This patch includes followup alterations to make the code match this policy