Author: Henrik Nordstrom <hno@squid-cache.org> Port from 2.7: maximum staleness limits
The default behaviour of Squid is to provide a stale copy (with Warning:
header) until an active response from the origin server causes the object
to be updated or garbage collection causes its removal.

The max_stale directive and refresh_pattern max-stale=N option allow admin
to set an upper limit on the objects age when serving stale responses.